readme update

+++ 14-binary_tree_balance.c
@@ -11,8 +11,8 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int left = binary_tree_height(tree->left);
-	int right = binary_tree_height(tree->right);
+	int left = (int) binary_tree_height(tree->left);
+	int right = (int) binary_tree_height(tree->right);
 
 	return (left - right);
 }
